+++ tests/Unit/DashboardTest.php
@@ -45,8 +45,11 @@ class DashboardTest extends TestUnitCase
         ]);
 
         $class = Dashboard::model(User::class);
+        $option = Dashboard::option('models.'.User::class);
 
         $this->assertEquals($class, 'MyCustomClass');
+        $this->assertEquals($option, 'MyCustomClass');
+        $this->assertEquals( Dashboard::option('random'), null);
     }
 
     protected function setUp()